item 추가
변경사항
>몬스터 별로 아이템 먹은 갯수를 저장
>아이템 생성시 bool 배열을 이용해 중복 체크
>지정된 범위에서만 이동하도록 한정
>pos를 Position으로 바꾸면서 생긴 사소한 오류 해결
>randerer에서 범위에 벗어나는 오브젝트는 draw하지 않도록 함
오류사항
>Sleep 함수가 프로그램은 정지시킴

+++ Screen/Screen/Bullets.h
@@ -122,7 +122,7 @@ public:
 			}
 		}
 
-		Borland::gotoxy(0, 3); printf("# of bullets = %2d ", container.size());		
+		Borland::gotoxy(0, 3); printf("# of bullets = %2d ", (int)container.size());
 		printf("%5s %20s\n", n_remaining_cool_time == 0 ? "ready" : " ", n_penetrable_cool_time == 0 ? "penCannonball ready" : " ");
 		Borland::gotoxy(0, 4); for (auto bullet : container) printf("%2.1f ", bullet->getDamagePower());
 	}
